website/src/test/java/de/benjaminborbe/website/util/UlWidgetUnitTest.java
package de.benjaminborbe.website.util;

import static org.junit.Assert.assertEquals;

import java.io.PrintWriter;
import java.io.StringWriter;

import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;

import org.easymock.EasyMock;
import org.junit.Test;

import de.benjaminborbe.html.api.HttpContext;

public class UlWidgetUnitTest {

	@Test
	public void testEmpty() throws Exception {
		final HttpServletRequest request = EasyMock.createMock(HttpServletRequest.class);
		EasyMock.replay(request);

		final StringWriter sw = new StringWriter();
		final PrintWriter writer = new PrintWriter(sw);

		final HttpServletResponse response = EasyMock.createMock(HttpServletResponse.class);
		EasyMock.expect(response.getWriter()).andReturn(writer).anyTimes();
		EasyMock.replay(response);

		final HttpContext context = EasyMock.createMock(HttpContext.class);
		EasyMock.replay(context);

		final UlWidget ul = new UlWidget();
		ul.render(request, response, context);
		assertEquals("<ul></ul>", sw.toString());
	}

	@Test
	public void testLi() throws Exception {
		final HttpServletRequest request = EasyMock.createMock(HttpServletRequest.class);
		EasyMock.replay(request);

		final StringWriter sw = new StringWriter();
		final PrintWriter writer = new PrintWriter(sw);

		final HttpServletResponse response = EasyMock.createMock(HttpServletResponse.class);
		EasyMock.expect(response.getWriter()).andReturn(writer).anyTimes();
		EasyMock.replay(response);

		final HttpContext context = EasyMock.createMock(HttpContext.class);
		EasyMock.replay(context);

		final UlWidget ul = new UlWidget();
		ul.add("l1");
		ul.add("l2");
		ul.render(request, response, context);
		assertEquals("<ul><li>l1</li><li>l2</li></ul>", sw.toString());
	}

	@Test
	public void testAddId() throws Exception {
		final HttpServletRequest request = EasyMock.createMock(HttpServletRequest.class);
		EasyMock.replay(request);

		final StringWriter sw = new StringWriter();
		final PrintWriter writer = new PrintWriter(sw);

		final HttpServletResponse response = EasyMock.createMock(HttpServletResponse.class);
		EasyMock.expect(response.getWriter()).andReturn(writer).anyTimes();
		EasyMock.replay(response);

		final HttpContext context = EasyMock.createMock(HttpContext.class);
		EasyMock.replay(context);

		final UlWidget ul = new UlWidget();
		ul.add("l1");
		ul.add("l2");
		ul.addAttribute("id", "testId");
		ul.render(request, response, context);
		assertEquals("<ul id=\"testId\"><li>l1</li><li>l2</li></ul>", sw.toString());
	}
}
